+++ Frameworks/PhunwareCameraPermission.xcframework/ios-arm64/PhunwareCameraPermission.framework/Headers/PhunwareCameraPermission.h
@@ -1,0 +1,19 @@
+//
+//  PhunwareCameraPermission.h
+//  PhunwareCameraPermission
+//
+//  Created by Kent Tu on 3/17/25.
+//  Copyright © 2025 Phunware, Inc. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+
+//! Project version number for PhunwareCameraPermission.
+FOUNDATION_EXPORT double PhunwareCameraPermissionVersionNumber;
+
+//! Project version string for PhunwareCameraPermission.
+FOUNDATION_EXPORT const unsigned char PhunwareCameraPermissionVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <PhunwareCameraPermission/PublicHeader.h>
+
+