+++ Frameworks/PhunwareMicrophonePermission.xcframework/ios-arm64_x86_64-simulator/PhunwareMicrophonePermission.framework/Headers/PhunwareMicrophonePermission.h
@@ -1,0 +1,19 @@
+//
+//  PhunwareMicrophonePermission.h
+//  PhunwareMicrophonePermission
+//
+//  Created by Kent Tu on 3/17/25.
+//  Copyright © 2025 Phunware, Inc. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+
+//! Project version number for PhunwareMicrophonePermission.
+FOUNDATION_EXPORT double PhunwareMicrophonePermissionVersionNumber;
+
+//! Project version string for PhunwareMicrophonePermission.
+FOUNDATION_EXPORT const unsigned char PhunwareMicrophonePermissionVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <PhunwareMicrophonePermission/PublicHeader.h>
+
+