+++ Frameworks/PhunwarePermissionCommon.xcframework/ios-arm64_x86_64-simulator/PhunwarePermissionCommon.framework/Headers/PhunwarePermissionCommon.h
@@ -1,0 +1,19 @@
+//
+//  PhunwarePermissionCommon.h
+//  PhunwarePermissionCommon
+//
+//  Created by Kent Tu on 3/17/25.
+//  Copyright © 2025 Phunware, Inc. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+
+//! Project version number for PhunwarePermissionCommon.
+FOUNDATION_EXPORT double PhunwarePermissionCommonVersionNumber;
+
+//! Project version string for PhunwarePermissionCommon.
+FOUNDATION_EXPORT const unsigned char PhunwarePermissionCommonVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <PhunwarePermissionCommon/PublicHeader.h>
+
+