+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.0][] - 2025-05-22
+
+### Added
+
+- Added support for dynamic loading
+- Added speech recognition permission 
+
 ## [1.4.0][] - 2024-07-13
 
 ### Added
@@ -56,6 +63,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
+[1.5.0]: https://github.com/phunware/artifact-permissions-ios/compare/1.4.0...1.5.0
 [1.4.0]: https://github.com/phunware/artifact-permissions-ios/compare/1.3.2...1.4.0
 [1.3.2]: https://github.com/phunware/artifact-permissions-ios/compare/1.3.1...1.3.2
 [1.3.1]: https://github.com/phunware/artifact-permissions-ios/compare/1.3.0...1.3.1